makeindexk: Avoid gcc warning on 64-bit systems
git-svn-id: svn://tug.org/texlive/trunk/Build/source@29973 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/makeindexk/mkind.h
@@ -506,7 +506,7 @@ extern NODE_PTR tail;
 
 extern int idx_dot;
 extern int idx_tt;
-extern size_t idx_gt;
+extern int idx_gt;
 extern int idx_et;
 extern int idx_dc;
